Add assessment type distinction for teacher and student-created assessments (#35)
* Add assessment type distinction for teacher and student-created assessments

* Add role validation for group members and enhance access control

* Add search filter to assessments and update group member role validation

* Enhance question access control to allow all authenticated users to access, edit, and create their own questions

* Add exclude_ids filter to question queries and parsing logic (#36)

* Feat/import export file (#37)

* Implement import/export functionality for questions with validation and error handling

* Add support for fill-in-the-blank, matching, and ordering question types in import/export functionality

* Add support for fill-in-the-blank, matching, and ordering question types in import/export functionality

* Implement group invite management with create, retrieve, delete, and join functionalities

* Enhance import/export service to support up to 10 options (A-J) for multiple choice questions

* Enhance grading permissions to allow assessment creators and group owners to grade answers

* Enhance question filtering and logging with Datadog integration

* Refactor assessment points to use float64 for improved precision

* Refactor assessment points to use float64 for improved precision

* Enhance soft delete functionality by adding DeletedAt field to multiple models

* Enhance auto-grading to log unanswered questions and include them in results

* Enhance attempt filtering by adding group ID and date range support (#41)

---------

+++ internal/events/notification_events.go
@@ -96,7 +96,7 @@ type AttemptGradedEvent struct {
 	StudentID       string    `json:"student_id"`
 	GradedAt        time.Time `json:"graded_at"`
 	Score           float64   `json:"score"`
-	MaxScore        int       `json:"max_score"`
+	MaxScore        float64   `json:"max_score"`
 	Percentage      float64   `json:"percentage"`
 	Passed          bool      `json:"passed"`
 	GraderID        string    `json:"grader_id"`
